refactor: extract types, constants, utils, and indicators from AdminCreateEventForm
Extracts 4 files from the 7832 LOC monolith (now 6836 LOC):
- admin-create-event-form-types.ts (30+ types/interfaces)
- admin-create-event-form-constants.ts (gas estimates, timeouts, options)
- admin-create-event-form-utils.ts (40+ pure helpers)
- admin-create-event-form-indicators.tsx (3 indicator components)

The main component body remains large (~6800 LOC) due to ~80 useState
calls with extreme interdependency. Further splitting requires
architectural refactoring beyond scope of this pass.
